refactor: move network status monitoring to a separate module
