+++ backend/src/network/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod status;
+pub(crate) mod types;
